Extract sick note controller security tests
