+++ src/test/java/org/synyx/urlaubsverwaltung/config/RestApiSecurityConfigIT.java
@@ -26,8 +26,6 @@ import static org.springframework.security.test.web.servlet.request.SecurityMock
 import static org.springframework.security.test.web.servlet.setup.SecurityMockMvcConfigurers.springSecurity;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-import static org.synyx.urlaubsverwaltung.person.Role.BOSS;
-import static org.synyx.urlaubsverwaltung.person.Role.DEPARTMENT_HEAD;
 import static org.synyx.urlaubsverwaltung.person.Role.OFFICE;
 import static org.synyx.urlaubsverwaltung.person.Role.SECOND_STAGE_AUTHORITY;
 
@@ -181,66 +179,6 @@ public class RestApiSecurityConfigIT {
     }
 
     @Test
-    public void getSickNotesWithoutBasicAuthIsUnauthorized() throws Exception {
-        final ResultActions resultActions = perform(get("/api/sicknotes"));
-
-        resultActions.andExpect(status().isUnauthorized());
-    }
-
-    @Test
-    public void getSickNotesWithBasicAuthIsOk() throws Exception {
-
-        createOfficePerson();
-
-        final LocalDateTime now = LocalDateTime.now();
-        final ResultActions resultActions = perform(get("/api/sicknotes")
-            .param("from", dtf.format(now))
-            .param("to", dtf.format(now.plusDays(5)))
-            .with(httpBasic("office", "secret")));
-
-        resultActions.andExpect(status().isOk());
-    }
-
-    @Test
-    public void getSicknotesWithNotPrivilegedUserIsForbidden() throws Exception {
-
-        createAuthenticatedPerson();
-
-        LocalDateTime now = LocalDateTime.now();
-        perform(get("/api/sicknotes")
-            .param("from", dtf.format(now))
-            .param("to", dtf.format(now.plusDays(5)))
-            .with(httpBasic("authenticated", "secret")))
-            .andExpect(status().isForbidden());
-    }
-
-    @Test
-    public void getSicknotesWithBossUserIsForbidden() throws Exception {
-
-        createBossPerson();
-
-        LocalDateTime now = LocalDateTime.now();
-        perform(get("/api/sicknotes")
-            .param("from", dtf.format(now))
-            .param("to", dtf.format(now.plusDays(5)))
-            .with(httpBasic("boss", "secret")))
-            .andExpect(status().isForbidden());
-    }
-
-    @Test
-    public void getSicknotesWithDepartmentHeadUserIsForbidden() throws Exception {
-
-        createDepartmentHeadPerson();
-
-        LocalDateTime now = LocalDateTime.now();
-        perform(get("/api/sicknotes")
-            .param("from", dtf.format(now))
-            .param("to", dtf.format(now.plusDays(5)))
-            .with(httpBasic("departmentHead", "secret")))
-            .andExpect(status().isForbidden());
-    }
-
-    @Test
     public void getSicknotesWithSecondStageUserIsForbidden() throws Exception {
 
         createSecondStagePerson();
@@ -344,20 +282,6 @@ public class RestApiSecurityConfigIT {
         secondStage.setPermissions(List.of(SECOND_STAGE_AUTHORITY));
         secondStage.setPassword("bc49b860775c4e6a813800fe827f093d40cd34a84134af9c6c67f5b68b0ccc43be73479103f8b714"); // secret
         personService.save(secondStage);
-    }
-
-    private void createDepartmentHeadPerson() {
-        final Person departmentHead = new Person("departmentHead", "Only", "Some", "");
-        departmentHead.setPermissions(List.of(DEPARTMENT_HEAD));
-        departmentHead.setPassword("bc49b860775c4e6a813800fe827f093d40cd34a84134af9c6c67f5b68b0ccc43be73479103f8b714"); // secret
-        personService.save(departmentHead);
-    }
-
-    private void createBossPerson() {
-        final Person boss = new Person("boss", "Only", "Some", "");
-        boss.setPermissions(List.of(BOSS));
-        boss.setPassword("bc49b860775c4e6a813800fe827f093d40cd34a84134af9c6c67f5b68b0ccc43be73479103f8b714"); // secret
-        personService.save(boss);
     }
 
     private ResultActions perform(MockHttpServletRequestBuilder builder) throws Exception {
